+++ CHANGELOG.md
@@ -10,6 +10,8 @@
    Где, `{ServiceName}` - имя сервиса. Например, **Bids** или **YandexAPIService**;
    `{MethodName}` - имя вызываемого метода API. Например, **Get** или **GetCampaignsParams**.
 
+ * **[BC BREAK]** Класс `YandexApiService` переименован в `YandexAPIService`, для соответствия с WSDL.
+
 ## 3.0.2 [commit logs](https://github.com/biplane/BiplaneYandexDirectBundle/compare/3.0.1...3.0.2)
 
  * Исправлена ошибка с отсутствием прав на запись, когда директория создается рекурсивно